refactor for testing
I hope this was the right thing to do. The code is kind of smelling to
me. I'm not sure why I structured the code to use events. I guess
because puppeteer uses events? Anyway, the code is the way it is ATM.

What I thought I was going to do for testing now that output exists
is generate an single output for each html file in test. The just
run each html file individual and expect the same output. I refactored
because of heavy part is starting and stopping puppeteer (chromium)
so I needed to make it so that only gets started once which is where
this code is ATM.